move ebool implementation to seperate files
