+++ test/node/dataset_pniio_bool_test.cpp
@@ -24,6 +24,7 @@
 //
 #include <gtest/gtest.h>
 #include <h5cpp/hdf5.hpp>
+#include <h5cpp/datatype/ebool.hpp>
 
 using namespace hdf5;
